requirements-dev.txt
# Core
torch>=2.0
torchvision
transformers>=4.46
pillow
pandas
pyyaml
tqdm
numpy

# Testing
pytest

# Jupyter
ipykernel
jupyter
matplotlib

# Config
pydantic>=2.0
